chose a different api becuase the futurama did not have a correct identifier. ALso, Triana helped and walked me through this. I got rid of context because it will created problems for testing
