+++ src/index.jsx
@@ -1,13 +1,10 @@
 import React from 'react';
 import { render } from 'react-dom';
 import App from './App';
-import { ShowProvider } from './context/ShowContext';
 
 render(
   <React.StrictMode>
-    <ShowProvider>
-      <App />
-    </ShowProvider>
+    <App />
   </React.StrictMode>,
   document.getElementById('root')
 );
