+++ src/views/Home.js
@@ -1,10 +1,10 @@
-import CharacterList from '../components/CharacterList';
+import QuoteList from '../components/QuoteList';
 
 export default function Home() {
   return (
     <div>
       Home Page is up G
-      <CharacterList />
+      <QuoteList />
     </div>
   );
 }
